+++ run_StorageVET.py
@@ -47,7 +47,7 @@ if __name__ == '__main__':
                                                  'analysis, dispatch, modelling, optimization, and valuation tool' +
                                                  '. Should be used with Python 3.8',
                                      epilog='Copyright 2022. Electric Power Research Institute (EPRI). ' +
-                                            'v1.2.1 All Rights Reserved.')
+                                            'v1.2.2 All Rights Reserved.')
     parser.add_argument('parameters_filename', type=str,
                         help='specify the filename of the CSV file defining the PARAMETERS dataframe')
     parser.add_argument('-v', '--verbose', action='store_true',